Rename to foodNameSearch for clear intent
